# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clawbolt",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/user/tools": {
      "get": {
        "summary": "Get Tool Config",
        "description": "Return the current tool configuration for the user.",
        "operationId": "get_tool_config_api_user_tools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolConfigResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Tool Config",
        "description": "Update tool configuration for the user.\n\nOnly domain-specific tools can be toggled. Attempts to disable\ncore tools are silently ignored.\n\nEach entry may include ``disabled_sub_tools`` to control individual\ntools within a factory group.",
        "operationId": "update_tool_config_api_user_tools_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubToolEntryResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "permission_level": {
            "type": "string",
            "title": "Permission Level",
            "default": "always"
          },
          "hidden_in_permissions": {
            "type": "boolean",
            "title": "Hidden In Permissions",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "enabled"
        ],
        "title": "SubToolEntryResponse"
      },
...
      "ToolConfigEntryResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "domain_group": {
            "type": "string",
            "title": "Domain Group",
            "default": ""
          },
          "domain_group_order": {
            "type": "integer",
            "title": "Domain Group Order",
            "default": 0
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured",
            "default": true
          },
          "auth_message": {
            "type": "string",
            "title": "Auth Message",
            "default": ""
          },
          "sub_tools": {
            "items": {
              "$ref": "#/components/schemas/SubToolEntryResponse"
            },
            "type": "array",
            "title": "Sub Tools"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "category",
          "enabled"
        ],
        "title": "ToolConfigEntryResponse"
      },
      "ToolConfigResponse": {
        "properties": {
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolConfigEntryResponse"
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "tools"
        ],
        "title": "ToolConfigResponse"
      },
      "ToolConfigUpdate": {
        "properties": {
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolConfigUpdateEntry"
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "tools"
        ],
        "title": "ToolConfigUpdate"
      },
      "ToolConfigUpdateEntry": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "disabled_sub_tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled Sub Tools"
          }
        },
        "type": "object",
        "required": [
          "name",
          "enabled"
        ],
        "title": "ToolConfigUpdateEntry"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}